frontend/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-docx==1.1.2
Pillow==10.4.0
python-multipart==0.0.9
python-dotenv==1.0.1

